+++ exports/mixins/vanilla-form-associated-mixin.js
@@ -154,7 +154,7 @@ export function VanillaFormAssociatedMixin(superclass) {
         if (!this.matches(":focus-within") && this.valueHasChanged) {
           this.hasInteracted = true
         }
-        runValidators(this)
+        this.runValidators()
       }
 
       get allValidators () {
@@ -233,7 +233,7 @@ export function VanillaFormAssociatedMixin(superclass) {
           }
         }
 
-        runValidators(this)
+        this.runValidators()
       }
 
       /**
@@ -249,7 +249,7 @@ export function VanillaFormAssociatedMixin(superclass) {
         this.value = this.defaultValue
         this.hasInteracted = false
         this.valueHasChanged = false
-        runValidators(this)
+        this.runValidators()
         this.setFormValue(this.defaultValue, this.defaultValue)
       }
 
@@ -261,7 +261,7 @@ export function VanillaFormAssociatedMixin(superclass) {
       formDisabledCallback(isDisabled) {
         this.disabled = isDisabled
         this.setValidity({})
-        runValidators(this) // Removes validations.
+        this.runValidators() // Removes validations.
       }
 
       /**
@@ -300,12 +300,12 @@ export function VanillaFormAssociatedMixin(superclass) {
       }
 
       reportValidity () {
-        runValidators(this)
+        this.runValidators()
         return this.internals.reportValidity()
       }
 
       checkValidity () {
-        runValidators(this)
+        this.runValidators()
         return this.internals.checkValidity()
       }
 
@@ -337,6 +337,57 @@ export function VanillaFormAssociatedMixin(superclass) {
       get form () {
         return this.internals.form
       }
+
+    /**
+    * @template {HTMLElement & ElementInternals & { formControl?: any, disabled?: boolean }} T
+    * @this {T}
+    */
+    runValidators () {
+      const element = this
+
+      if (element.disabled || element.getAttribute("disabled")) {
+        element.setValidity({})
+        // We don't run validators on disabled elements to be inline with native HTMLElements.
+        // https://codepen.io/paramagicdev/pen/PoLogeL
+        return
+      }
+
+      const validators = /** @type {{allValidators?: Array<import("../types.js").Validator>}} */ (/** @type {unknown} */ (element)).allValidators
+
+      if (!validators) {
+        element.setValidity({})
+        return
+      }
+
+      const flags = {
+        customError: element.validity.customError
+      }
+
+      const formControl = element.formControl || undefined
+
+      let finalMessage = ""
+
+      for (const validator of validators) {
+        const { isValid, message, invalidKeys } = validator.checkValidity(element)
+
+        if (isValid) { continue }
+
+        if (!finalMessage) {
+          finalMessage = message
+        }
+
+        if (invalidKeys?.length >= 0) {
+          // @ts-expect-error
+          invalidKeys.forEach((str) => flags[str] = true)
+        }
+      }
+
+      // This is a workaround for preserving custom errors
+      if (!finalMessage) {
+        finalMessage = element.validationMessage
+      }
+      element.setValidity(flags, finalMessage, formControl)
+    }
     }
   )
 }